Delete jobsheet4 and renamed to minggu4, update sort java
