+++ 1941720169/AlgoStruDat/src/Minggu5/Mahasiswa.java
@@ -1,0 +1,5 @@
+package Minggu5;
+
+public class Mahasiswa {
+    
+}
