+++ 1941720169/AlgoStruDat/src/TugasMinggu3/NilaiModif.java
@@ -1,5 +1,5 @@
 
-package Tugas;
+package TugasMinggu3;
 
 public class NilaiModif {
     public String namaMhs;
