+++ 1941720169/Sort Java/src/sort/Bubble.java
@@ -12,6 +12,7 @@ public class Bubble {
                 k = j + 1;
                 if (arr[j] > arr[k]) {
                     swap(j, k, arr);
+                    System.out.println("j: "+j+" | "+"k: "+k); 
                 }
             }
             print(arr);
